--nbsmodule  * resurrect old module * address compatibility with later openmrs version * modify for major atd and sockethl7listener module changes. * add new nbs_alert_status table * modify process for construction of hl7 message by ResponseGenerator based on alert state * modify hl7 message construction pertaining to D4D requirements * change interval between sending messages to be configurable due to large hl7 messages containing encoded tiffs. * use of new module ChirdlUtil for standard util methods.

+++ src/org/openmrs/module/nbsmodule/impl/NBSServiceImpl.java
@@ -1,7 +1,12 @@
 package org.openmrs.module.nbsmodule.impl;
 
 import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.Date;
@@ -16,6 +21,7 @@ import org.apache.commons.logging.LogFactory;
 import org.hibernate.HibernateException;
 import org.openmrs.Encounter;
 import org.openmrs.Form;
+import org.openmrs.Obs;
 import org.openmrs.Patient;
 import org.openmrs.PatientIdentifier;
 import org.openmrs.api.APIException;
@@ -26,14 +32,16 @@ import org.openmrs.api.context.Context;
 import org.openmrs.logic.result.Result;
 import org.openmrs.module.atd.TeleformFileMonitor;
 import org.openmrs.module.atd.hibernateBeans.FormAttributeValue;
+import org.openmrs.module.atd.hibernateBeans.FormInstance;
+import org.openmrs.module.atd.hibernateBeans.Session;
 import org.openmrs.module.atd.service.ATDService;
 import org.openmrs.module.dss.hibernateBeans.Rule;
 import org.openmrs.module.dss.service.DssService;
-import org.openmrs.module.dss.util.IOUtil;
 import org.openmrs.module.nbsmodule.NBSModuleResponse;
 import org.openmrs.module.nbsmodule.NBSService;
 import org.openmrs.module.nbsmodule.db.NBSModuleDAO;
 import org.openmrs.module.nbsmodule.util.Util;
+import org.openmrs.module.nbsmodule.util.Base64;
 import org.openmrs.module.sockethl7listener.HL7MessageConstructor;
 
 /**
@@ -52,9 +60,6 @@ public class NBSServiceImpl implements NBSService {
 	
 	public NBSServiceImpl() {
 		log.error("**********NBSService Constructor called!!!!!!!! ");
-		
-		
-	   
 	}
 	
 	private NBSModuleDAO getNBSModuleDAO() {
@@ -84,7 +89,7 @@ public class NBSServiceImpl implements NBSService {
 	 * @throws APIException
 	 */
 	public NBSModuleResponse getNBSModuleResponse(String pid ) throws APIException {
-		String patientID;
+		String patientID = null;
 		int index = pid.indexOf("=");
 		
 		if (index == -1){
@@ -186,6 +191,13 @@ public class NBSServiceImpl implements NBSService {
 		
 		ATDService atdService = Context.getService(ATDService.class);
 		List<Form> forms = formService.getForms(formName, false);
+		AdministrationService adminService = Context.getAdministrationService();
+		String rulePackagePrefix = adminService.getGlobalProperty("dss.rulePackagePrefix");
+		List<Session> sessions = atdService.getSessionsByEncounter(enc.getEncounterId());
+		Session session = null;
+		if (sessions.size()>0){
+			session = sessions.get(0);
+		}
 		if (forms.isEmpty())
 		{
 			//TODO: Throw exception
@@ -198,9 +210,15 @@ public class NBSServiceImpl implements NBSService {
 		result = atdService.evaluateRule(tokenName,
 				patient, new HashMap<String,Object>(), null);
 		
-		if(result != null && !result.exists())
+		if(result == null || !result.exists())
 		{	
 			return;
+		}
+		
+		String mergeDirectory = null;
+		FormAttributeValue val = atdService.getFormAttributeValue(formId, "defaultMergeDirectory", 1,1 );
+		if (val != null) {
+			mergeDirectory = val.getValue(); 
 		}
 		
 		Map<String, Object> parameters = null;
@@ -253,11 +271,13 @@ public class NBSServiceImpl implements NBSService {
 			for(int i = 0; i < result.size(); i++)
 			{
 				NBSModuleResponse response = new NBSModuleResponse();
+				Integer providerId = enc.getProvider().getUserId();
 				response.setPatient(patient);
 				
 				if(ruleFieldname.equalsIgnoreCase("provider"))
 				{
 					response.setProvider(result.get(i).toString());
+					response.setProviderId(providerId);
 				}
 				else {
 					response.setProvider(enc.getProvider().toString());
@@ -265,32 +285,49 @@ public class NBSServiceImpl implements NBSService {
 				}
 				
 				
-				response.setDatestamp(new Date());
+				response.setDateCreated(new Date());
 				response.setFormId(formId);
 				
-				
-				if (!getNBSModuleDAO().duplicateAlert(response))
-				{
+				List<NBSModuleResponse> existingAlerts = getNBSAlerts(patient, formId, 
+						providerId, enc, new Date(), null);
+				//Don't send to physician multiple times per day.
+				if (existingAlerts.size() > 0){
+					return;
+				}
 					
-				int formInstanceId = atdService.addFormInstance(formId).getFormInstanceId();
+				FormInstance formInstance = atdService.addFormInstance(formId, enc.getLocation().getLocationId() );
+				if (formInstance != null){
+					Integer formInstanceId = formInstance.getFormInstanceId();
 				
-				FormAttributeValue val = atdService.getFormAttributeValue(formId, "defaultMergeDirectory");
-				String mergeDirectory = val.getValue(); 
-				String mergeFilename = mergeDirectory + "//" + formInstanceId + ".xml";
+					String mergeFilename = mergeDirectory + "//" + formInstance + ".xml";
 										
-				OutputStream output = new FileOutputStream(mergeFilename);
-				parameters.put(ruleFieldname, result.get(i).toString());
+					OutputStream output = new FileOutputStream(mergeFilename);
+					parameters.put(ruleFieldname, result.get(i).toString());
 				
-				if(atdService.produce(patient, formInstanceId,output,formId,
-						enc.getEncounterId(),parameters, null, false)) {
-					response.setEncounter(enc);
-					response.setMrn(patient.getPatientIdentifier().getIdentifier());
-					response.setAlertId(formInstanceId);
-					createNBSModuleResponse(response);
+					
+					if (atdService.produce(patient, formInstance, output, 
+							enc.getEncounterId(),
+							parameters,
+							rulePackagePrefix,
+							false ,1 ,
+							session.getSessionId()))
+					{
+						response.setEncounter(enc);
+						response.setMrn(patient.getPatientIdentifier().getIdentifier());
+						response.setFormInstanceId(formInstance.getFormInstanceId());
+						Integer status = getAlertStatusIdByName("waiting_for_export");
+						if (status == null){
+							status = 1;
+						}
+						response.setStatus(status);
+						createNBSModuleResponse(response);
+					}
+
 				}
 				
-				TeleformFileMonitor.addToPendingStatesWithFilename(formId, formInstanceId, mergeDirectory + formInstanceId +" .19");
-				}
+				TeleformFileMonitor.addToPendingStatesWithFilename(formInstance, mergeDirectory 
+						+ formInstance.getFormInstanceId() +" .19");
+				
 			}
 		}
 		else
@@ -298,17 +335,22 @@ public class NBSServiceImpl implements NBSService {
 			NBSModuleResponse response = new NBSModuleResponse();
 			response.setPatient(patient);
 			response.setProviderId(enc.getProvider().getUserId());
-			response.setDatestamp(new Date());
+			response.setDateCreated(new Date());
 			response.setFormId(formId);
 			
-			if (!getNBSModuleDAO().duplicateAlert(response))
-			{
-				
-			int formInstanceId = atdService.addFormInstance(formId).getFormInstanceId();
 			
-			org.openmrs.module.atd.hibernateBeans.FormAttributeValue val = atdService.getFormAttributeValue(formId, "defaultMergeDirectory");
-			String mergeDirectory = val.getValue(); 
-			String mergeFilename = mergeDirectory + "//" + formInstanceId + ".xml";
+			
+			List<NBSModuleResponse> existingAlerts = getNBSAlerts(patient, formId, 
+					enc.getProvider().getUserId(), enc,  new Date(), null);
+			//Don't send to physician multiple times per day.
+			if (existingAlerts.size() > 0){
+				return;
+			}
+				
+			FormInstance formInstance = atdService.addFormInstance(formId, enc.getLocation().getLocationId());
+			Integer formInstanceId = formInstance.getFormInstanceId();
+			
+			String mergeFilename = mergeDirectory + "//" + formInstance + ".xml";
 									
 			OutputStream output = new FileOutputStream(mergeFilename);
 			
@@ -323,152 +365,60 @@ public class NBSServiceImpl implements NBSService {
 			ruleFieldname = "paragraphText";
 			parameters.put(ruleFieldname, result.get(0).toString());
 			
-			if(atdService.produce(patient, formInstanceId,output,formId,
-					enc.getEncounterId(),parameters, null, false)) {
+			if (atdService.produce(patient, formInstance, output, 
+					enc.getEncounterId(),
+					parameters,
+					rulePackagePrefix,
+					false ,1 ,
+					session.getSessionId()))
+			{
 
 				response.setEncounter(enc);
 				response.setProvider(enc.getProvider().toString());
-
 				response.setMrn(patient.getPatientIdentifier().getIdentifier());
-				
-				response.setAlertId(formInstanceId);
+				response.setFormInstanceId(formInstance.getFormInstanceId());
+				Integer status = getAlertStatusIdByName("waiting_for_export");
+				if (status == null){
+					status = 1;
+				}
+				response.setStatus(status);
 				createNBSModuleResponse(response);
 			}
 			
-			TeleformFileMonitor.addToPendingStatesWithFilename(formId, formInstanceId, mergeDirectory + formInstanceId +" .19");
-			}
+			TeleformFileMonitor.addToPendingStatesWithFilename( formInstance , mergeDirectory + formInstanceId +" .19");
+			
 		}
 		
 	}
-	
 	
 	
 	/* (non-Javadoc)
 	 * @see org.openmrs.module.nbsmodule.NBSService#getAlerts(org.openmrs.Patient, java.util.Properties)
 	 */
-	public List<String>  getAlerts(Patient resultPatient, Properties prop)
-	{
-		Integer pid = null;
-		String message = "";
-		if (resultPatient!= null){
-			pid = resultPatient.getPatientId();
-		}
-		//if pid is null, get all alerts not yet processed
-		List<NBSModuleResponse> nbsAlerts = getNBSModuleDAO().getNBSAlertsByPatient(pid);
-		List<String> messages = new ArrayList<String>();
+	public List<NBSModuleResponse>  getNBSAlerts(Patient resultPatient, Integer formId, Integer providerId, Encounter encounter, Date date, Integer status){
+	
+		List<NBSModuleResponse> nbsAlerts = getNBSModuleDAO().getNBSAlerts(resultPatient,formId, providerId, encounter, date, status);
+		return nbsAlerts;
+	}
 		
-		try
-		{
-			for(NBSModuleResponse currAlert:nbsAlerts)
-			{
-				int encounter_id = 0;
-				int nbs_alert_id = 0;
-				int alert_id = 0;
-				String mrn = "";
-				int form_id = 0;
-				if (currAlert.getEncounter().getEncounterId() != null)
-				{
-					encounter_id = currAlert.getEncounter().getEncounterId();
-				} else
-				{
-					log.error("ERROR: Null value for encounter_id in nbs_alert table ");
-					break;
-				}
-				if (currAlert.getMrn() != null)
-				{
-					mrn = currAlert.getMrn();
-				} else
-				{
-					// continue with MRN?
-					log.error("ERROR: Null value for mrn in nbs_alert table.");
-				}
-
-				nbs_alert_id = currAlert.getNbsalertId();
-
-				if (currAlert.getAlertId() != null)
-				{
-					alert_id = currAlert.getAlertId();
-				}
-				if (currAlert.getFormId() != null)
-				{
-					form_id = currAlert.getFormId();
-				}
-				ATDService atdService = Context.getService(ATDService.class);
-				FormAttributeValue value = atdService.getFormAttributeValue(form_id, "defaultImageDirectory");
-				value.getValue();
-
-				HL7MessageConstructor hl7mc = new HL7MessageConstructor(prop);
-				message = hl7mc.createHl7Message(encounter_id, mrn, value.getValue(),
-				alert_id);
-				if (message != null)
-				{
-					currAlert.setStatus(1);
-					getNBSModuleDAO().updateNBSModuleResponse(currAlert);
-					messages.add(message);
-					saveMessage(message, encounter_id);
-				}
-				
-				
-			
-
+	public List<NBSModuleResponse>  captureNBSAlerts(Patient resultPatient, Integer formId, Integer providerId, Encounter encounter, Date date){
+		NBSService nbsService = Context.getService(NBSService.class);
+		Integer status = getAlertStatusIdByName("waiting_for_export");
+		List<NBSModuleResponse> nbsAlerts = getNBSModuleDAO().getNBSAlerts(resultPatient,formId, providerId, null, null, status );
+		for (NBSModuleResponse alert : nbsAlerts){
+			Integer statusId = getAlertStatusIdByName("processing_alert");
+			if (status == null){
+				status = 1;
 			}
-
-	        	
-		} catch (HibernateException e)
-		{
-			e.printStackTrace();
-		} catch (Exception e){
-			Util.getStackTrace(e);
+			alert.setStatus(statusId);
+			nbsService.updateNBSModuleResponse(alert);
 		}
-		return messages;
+		return nbsAlerts;
 	}
 	
-	/**
-	 * Saves message string to archive directory
-	 * @param message
-	 * @param encid
-	 */
-	public void saveMessage( String message, Integer encid){
-		AdministrationService adminService = Context.getAdministrationService();
-		EncounterService es = Context.getService(EncounterService.class);
-		String archiveDir = "";
-		String filename = "";
-		try {
-			Encounter enc = es.getEncounter(encid);
-			Patient patient = new Patient();
-			patient = enc.getPatient();
-			PatientIdentifier pi = patient.getPatientIdentifier();
-			String mrn = "";
-		
-		
-			if (pi != null) mrn = pi.getIdentifier();
-			 filename =  Util.archiveStamp() + "_"+ mrn + ".hl7";
-		
-			archiveDir = IOUtil.formatDirectoryName(adminService
-				.getGlobalProperty("nbsmodule.outboundHl7ArchiveDirectory"));
-		
-
-			FileOutputStream archiveFile = null;
-		
-			archiveFile = new FileOutputStream(
-					archiveDir + "/" + filename);
-		
-			if (archiveDir != null && archiveFile != null)
-			{
-			
-				ByteArrayInputStream messageStream = new ByteArrayInputStream(
-						message.getBytes());
-				IOUtil.bufferedReadWrite(messageStream, archiveFile);
-				archiveFile.flush();
-				archiveFile.close();
-				
-			}
-			log.error("Couldn't find file: "+archiveDir + "/" + filename);
-			
-		} catch (Exception e){
-			log.error(Util.getStackTrace(e));
-		}
-		return;
-		
+	public Integer getAlertStatusIdByName(String name){
+		return getNBSModuleDAO().getAlertStatusIdByName(name);
 	}
+
+	
 }
